+++ odtp/dashboard/page_run/run.py
@@ -1,154 +1,228 @@
-import asyncio
-import logging
-import os.path
-import shlex
+import json
+import os
 import sys
+import asyncio
+from pprint import pprint
+import odtp.dashboard.page_run.helpers as rh
+import os.path
+import shutil
+import shlex
+from slugify import slugify
+from nicegui import ui, app
+import odtp.mongodb.db as db
+import odtp.dashboard.utils.helpers as helpers
+import odtp.dashboard.utils.ui_theme as ui_theme
+import odtp.dashboard.utils.storage as storage
+import odtp.dashboard.page_run.folder as folder
+import odtp.dashboard.utils.validators as validators
+
 
 from nicegui import ui
-import odtp.helpers.settings as config
-import odtp.dashboard.page_run.helpers as rh
-import odtp.dashboard.page_run.folder as folder
 
-log = logging.getLogger(__name__)
-log.setLevel(logging.INFO)
-log.addHandler(config.get_command_log_handler())
+class ExecutionRunForm:
+    def __init__(self, digital_twin_id, execution_id, workdir, dialog, result):
+        """init form"""
+        self.workdir = workdir
+        self.execution_id = execution_id
+        self.digital_twin_id = digital_twin_id
+        self.execution_path = None
+        self.dialog = dialog
+        self.result = result
+        self.folder_status = None
+        self.get_execution()
+        self.check_folder_status()
+        self.build_form()
 
+    def get_execution(self):
+        if not self.execution_id:
+            return
+        self.execution = helpers.build_execution_with_steps(self.execution_id)
+        if self.execution.get("execution_path"):
+            self.execution_path = self.execution.get("execution_path")
+        self.ui_execution_info.refresh()
 
-def ui_prepare_execution(dialog, result, current_run, folder_status):
-    stepper = current_run.get("stepper")
-    if stepper and rh.STEPPERS.index(stepper) != rh.STEPPER_PREPARE_EXECUTION:
-        return
-    with ui.row().classes("w-full"):
-        ui.markdown(
-            f"""
-            ###### Prepare the Execution: 
-
-            - clone github repo
-            - build Docker images 
-            - create folder structure
-            """
+    @ui.refreshable
+    def ui_add_project_path(self):
+        if not self.execution:
+            return
+        if self.execution_path:
+            return
+        project_folder_input = ui.input(
+            value=self.execution_path,
+            label="Project folder name",
+            placeholder="execution",
+            validation={
+                f"Project folder already exists and is not empty": lambda value: validators.validate_folder_does_not_exist(
+                    value, self.workdir
+                )
+            },
         )
-        if folder_status == rh.FOLDER_PREPARED:
-            with ui.row().classes("w-full"):
-                rh.ui_display_folder_status(folder_status)
-        if folder_status == rh.FOLDER_EMPTY:
-            execution = current_run["execution"]
-            project_path = current_run["project_path"]
+        ui.button(
+            "Create new project folder",
+            on_click=lambda: self.create_folder(project_folder_input),
+            icon="add",
+        ).props("flat ")
+
+    def check_folder_status(self):
+        self.folder_status = rh.get_folder_status(self.execution_id, self.execution_path)
+
+    def create_folder(self, folder_name_input):
+        if self.execution_path:
+            return
+        if (
+            not folder_name_input.validate()
+        ):
+            ui.notify(
+                "Project folder already exists and is not empty", type="negative"
+            )
+            return
+        folder_name = folder_name_input.value
+        project_path = os.path.join(self.workdir, folder_name)
+        if os.path.exists(project_path):
+            shutil.rmtree(project_path)
+        os.makedirs(project_path)
+        self.execution_path = project_path
+        db.set_execution_path(self.execution_id, execution_path=project_path)
+        ui.notify(
+            f"project directory {project_path} has been created and set as project directory",
+            type="positive",
+        )
+        self.ui_add_project_path.refresh()
+
+    @ui.refreshable
+    def ui_show_project_folder(self):
+        if not self.execution_path:
+            return
+        with ui.row().classes("w-full"):
+            cli_tree_command = f"tree {self.execution_path}"
+            with ui.grid(columns=1):
+                #with ui.row().classes("w-full"):
+                #    ui.label(cli_tree_command).classes("font-mono w-full")
+                with ui.row().classes("w-full"):
+                    ui.button(
+                        "Show execution path",
+                        on_click=lambda: self.run_command(cli_tree_command),
+                        icon="folder",
+                    ).props("flat")
+
+    @ui.refreshable
+    def ui_execution_info(self):
+        if not self.execution_id:
+            return
+        with ui.expansion("execution info", icon="info"):
+            ui.label(f"Run for execution {self.execution_id}")
+            ui.label(f"created: {self.execution['createdAt']}")
+            ui.label(f"inputs: {self.execution['inputs']}")
+            ui.label(f"outputs: {self.execution['outputs']}")
+            ui.label(f"ports: {self.execution['ports']}")
+            ui.label(f"steps: {self.execution['steps']}")
+            ui.label(f"version_tags: {self.execution['version_tags']}")
+            ui.label(f"versions: {self.execution['versions']}")
+            ui.label(f"path: {self.execution.get('execution_path')}")
+
+    @ui.refreshable
+    def build_form(self):
+        """render form elements"""
+        self.ui_execution_info()
+        self.ui_add_project_path()
+        self.ui_display_folder_status()
+        self.ui_prepare_execution()
+        self.ui_show_project_folder()
+        self.ui_run_execution()
+
+    def ui_prepare_execution(self):
+        if not self.execution_path:
+            return
+        if self.folder_status == rh.FOLDER_PREPARED:
+            return
+        with ui.row().classes("w-full"):
             cli_prepare_command = rh.build_cli_command(
                 cmd="prepare",
-                execution_id=execution["execution_id"],
-                project_path=project_path,
-            )        
+                execution_id=self.execution_id,
+                project_path=self.execution_path,
+            )
             with ui.grid(columns=1):
-                with ui.row().classes("w-full"):
-                    ui.label(cli_prepare_command).classes("font-mono w-full")
+                #with ui.row().classes("w-full"):
+                #    ui.label(cli_prepare_command).classes("font-mono w-full")
                 with ui.row().classes("w-full"):
                     ui.button(
                         "Prepare execution",
-                        on_click=lambda: run_command(cli_prepare_command, dialog, result),
+                        on_click=lambda: self.run_command(cli_prepare_command),
                         icon="folder",
                     ).props("no-caps")
-    rh.ui_next_back(current_run, ready_for_next=True)
 
 
-def ui_run_execution(dialog, result, current_run, folder_status):
-    stepper = current_run.get("stepper")
-    if stepper and rh.STEPPERS.index(stepper) != rh.STEPPER_RUN_EXECUTION:
-        return
-    with ui.row().classes("w-full"):
-        ui.markdown(
-            f"""
-            ###### Run the Execution: 
-
-            - Run docker images as containers
-            - write output 
-            """
-        )
-    execution = current_run["execution"]
-    project_path = current_run["project_path"]
-    secret_files = current_run["secret_files"]
-    if folder_status >= rh.FOLDER_PREPARED:
-        cli_run_command = rh.build_cli_command(
-            cmd="run",
-            secret_files=secret_files,
-            execution_id=execution["execution_id"],
-            project_path=project_path,
-        )
-        cli_log_commands = []
-        for i, _ in enumerate(execution["versions"]):
-            cli_log_commands. append(rh.build_cli_command(
-                cmd="streamlogs",
-                project_path=project_path,
-                step_nr=str(i)
-            ))
+    def ui_run_execution(self):
+        if not self.execution_path:
+            return
+        if self.folder_status != rh.FOLDER_PREPARED:
+            return
         with ui.row().classes("w-full"):
-            ui.label(cli_run_command).classes("font-mono")
-        with ui.row().classes("w-full"):
-            ui.button(
-                "Run execution",
-                on_click=lambda: submit_command(cli_run_command),
-                icon="rocket",
-            ).props("no-caps")
-        with ui.row().classes("w-full"):
-            ui.icon("warning").classes("text-lg text-yellow")
-            ui.label(
-                """The logs for a step only become available once it is running.
-                So if they are not available right away, they may be so when you 
-                click the button again.
-                """
-            )            
-        for i, cli_log_command in enumerate(cli_log_commands):    
-            with ui.row().classes("w-full"):
-                ui.button(
-                    f"show logs for step {i}",
-                    on_click=lambda cli_log_command=cli_log_command: run_command(cli_log_command, dialog, result),
-                    icon="info",
-                ).props("no-caps")
-    rh.ui_next_back(current_run, ready_for_next=False)
+            cli_run_command = rh.build_cli_command(
+                cmd="run",
+                execution_id=self.execution_id,
+                project_path=self.execution_path,
+            )
+            with ui.grid(columns=1):
+                with ui.row().classes("w-full"):
+                    ui.label(cli_run_command).classes("font-mono w-full")
+                with ui.row().classes("w-full"):
+                    ui.button(
+                        "Run execution",
+                        on_click=lambda: self.run_command(cli_run_command),
+                        icon="folder",
+                    ).props("no-caps")
 
+    @ui.refreshable
+    def ui_display_folder_status(self):
+        if not self.execution_path:
+            return
+        with ui.grid(columns='1fr 7fr'):
+            if self.folder_status == rh.FOLDER_EMPTY:
+                ui.icon("check").classes("text-teal text-lg")
+                ui.label("Project folder for the execution run has been selected").classes("text-teal")
+            elif self.folder_status == rh.FOLDER_NOT_SET:
+                ui.icon("clear").classes("text-red text-lg")
+                ui.label("Project folder missing: please select one").classes("text-red")
+            elif self.folder_status == rh.FOLDER_DOES_NOT_MATCH:
+                ui.icon("clear").classes("text-red text-lg")
+                ui.label("The project folder structure does not match the steps of the execution: choose an empty project folder or create a new project folder").classes("text-red")
+            elif self.folder_status == rh.FOLDER_PREPARED:
+                ui.icon("check").classes("text-teal text-lg")
+                ui.label("Project folder for the execution run has been selected").classes("text-teal")
+                ui.icon("check").classes("text-teal text-lg")
+                ui.label("The execution has been prepared").classes("text-teal")
+            elif self.folder_status == rh.FOLDER_HAS_OUTPUT:
+                ui.icon("check").classes("text-teal text-lg")
+                ui.label("Project folder for the execution run has been selected").classes("text-teal")
+                ui.icon("check").classes("text-teal text-lg")
+                ui.label("The execution has been prepared").classes("text-teal")
+                ui.icon("check").classes("text-teal text-lg")
+                ui.label("The execution has been run").classes("text-teal")
 
-async def submit_command(command):
-    try:
-        log.info(command)
-        log.info(shlex.split(command, posix="win" not in sys.platform.lower()))
-        process = await asyncio.create_subprocess_exec(
-            *shlex.split(command, posix="win" not in sys.platform.lower()),
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.STDOUT,
-            cwd=os.path.dirname(os.path.abspath(__file__)),
-        )
-        ui.notify(f"{command} has been submitted as {process}. Click on 'Show logs to view the progress", type="positive")
-    except Exception as e:
-        log.exception(f"run command failed with Exception {e}")
-
-
-async def run_command(command: str, dialog, result) -> None:
-    """Run a command in the background and display the output in the pre-created dialog.
-    This function has been copied from nicegui examples.
-    """
-    try:
-        dialog.open()
-        result.content = "... loading"
-        log.info(command)
-        log.info(shlex.split(command, posix="win" not in sys.platform.lower()))
-        process = await asyncio.create_subprocess_exec(
-            *shlex.split(command, posix="win" not in sys.platform.lower()),
-            stdout=asyncio.subprocess.PIPE,
-            stderr=asyncio.subprocess.STDOUT,
-            cwd=os.path.dirname(os.path.abspath(__file__)),
-        )
-        # NOTE we need to read the output in chunks, otherwise the process will block
-        output = ""
-        while True:
-            new = await process.stdout.read(100000)
-            if not new:
-                break
-            output += new.decode()
-            # NOTE the content of the markdown element is replaced every time we have new output
-            result.content = f"```\n{output}\n```"
-    except Exception as e:
-        log.exception(f"run command failed with Exception {e}")
-    else:        
-        from odtp.dashboard.page_run.main import ui_workarea, ui_stepper    
-        ui_workarea.refresh()
-        ui_stepper.refresh()
+    async def run_command(self, command: str):
+        try:
+            self.dialog.open()
+            self.result.content = "... loading"
+            process = await asyncio.create_subprocess_exec(
+                *shlex.split(command, posix="win" not in sys.platform.lower()),
+                stdout=asyncio.subprocess.PIPE,
+                stderr=asyncio.subprocess.STDOUT,
+                cwd=os.path.dirname(os.path.abspath(__file__)),
+            )
+            # NOTE we need to read the output in chunks, otherwise the process will block
+            output = ""
+            while True:
+                new = await process.stdout.read(100000)
+                if not new:
+                    break
+                output += new.decode()
+                # NOTE the content of the markdown element is replaced every time we have new output
+                self.result.content = f"```\n{output}\n```"
+        except Exception as e:
+            print(f"run command failed with Exception {e}")
+        else:
+            self.check_folder_status()
+            self.ui_display_folder_status.refresh()
+            self.ui_execution_info.refresh()
+            self.build_form.refresh()
